+++ boiler/commands/settings/statusOptions/index.js
@@ -1,0 +1,5 @@
+module.exports = [
+  require('./add'),
+  require('./del'),
+  require('./view')
+]
